use sha256
BLAKE3 has a weird problem that it produces different hashes for the same input when run as native Rust vs in Wasm. It seems to have something to do with the binaryen optimizer. I also can't reproduce the issue anywhere out of this repo. SHA256 doesn't have this problem so we use it for now. See: https://twitter.com/larry0x/status/1744954350277894205